UWRITE/requirements.txt
requests==2.10.0
PyPDF2==1.26.0
svglib==0.9.2
reportlab==3.5.23
PyYAML==5.1.2